Refactor DisplaySize to Bounds, fix various bugs, etc
DisplaySize used to only take a width and height to allow the widget to only be concerned with its dimensions, but the reality is the rest of the application is greatly concerned with Rects and where to draw things. So this is part 1 of a multipart commit series where I try to focus more on space used by widgets.

In the future, I want to allow Widgets to receive a `space Rect` in all signatures to decide how they want to render to take up available space, etc, instead of the parent calling `Draw(child.Bounds(space))` which is more restrictive and redundant.